+++ src/main/java/com/gistpetition/api/user/application/SessionLoginService.java
@@ -1,12 +1,12 @@
 package com.gistpetition.api.user.application;
 
-import com.gistpetition.api.common.password.Encoder;
 import com.gistpetition.api.exception.user.NoSuchUserException;
 import com.gistpetition.api.exception.user.NotMatchedPasswordException;
 import com.gistpetition.api.user.domain.SimpleUser;
 import com.gistpetition.api.user.domain.User;
 import com.gistpetition.api.user.domain.UserRepository;
 import com.gistpetition.api.user.dto.request.SignInRequest;
+import com.gistpetition.api.utils.password.Encoder;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
