+++ src/main/java/com/gistpetition/api/utils/emailsender/EmailSender.java
@@ -1,4 +1,4 @@
-package com.gistpetition.api.emailsender;
+package com.gistpetition.api.utils.emailsender;
 
 public interface EmailSender {
     void send(String to, String subject, String content);
