+++ src/main/java/com/gistpetition/api/utils/emailsender/FakeEmailSender.java
@@ -1,4 +1,4 @@
-package com.gistpetition.api.emailsender;
+package com.gistpetition.api.utils.emailsender;
 
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Service;
