+++ src/test/java/com/gistpetition/api/ServiceTest.java
@@ -1,6 +1,6 @@
 package com.gistpetition.api;
 
-import com.gistpetition.api.emailsender.EmailSender;
+import com.gistpetition.api.utils.emailsender.EmailSender;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
