+++ src/test/java/com/gistpetition/api/user/LoginServiceTest.java
@@ -1,6 +1,5 @@
 package com.gistpetition.api.user;
 
-import com.gistpetition.api.common.password.Encoder;
 import com.gistpetition.api.exception.user.NoSuchUserException;
 import com.gistpetition.api.exception.user.NotMatchedPasswordException;
 import com.gistpetition.api.user.application.LoginService;
@@ -9,6 +8,7 @@ import com.gistpetition.api.user.domain.User;
 import com.gistpetition.api.user.domain.UserRepository;
 import com.gistpetition.api.user.domain.UserRole;
 import com.gistpetition.api.user.dto.request.SignInRequest;
+import com.gistpetition.api.utils.password.Encoder;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Test;
